feat#7: 파일 정리
