+++ src/component/MainWrapper/interface.d.ts
@@ -1,0 +1,3 @@
+export interface HeaderProp {
+  isHeader: boolean;
+}
