+++ src/hook/useCategory/interface.d.ts
@@ -1,0 +1,7 @@
+import { ReactElement } from "react";
+
+export interface CategoryItemProps {
+  children: ReactElement | string;
+  action: () => void;
+  toggle?: boolean;
+}
