+++ src/hook/useDrawer/inderface.d.ts
@@ -1,0 +1,3 @@
+export interface DrawHeaderProps {
+  drawerBleeding: number;
+}
